+++ src/module/active-effects/effects.js
@@ -3,6 +3,9 @@
  * @param {MouseEvent} event      The left-click event on the effect control
  * @param {Actor|Item} owner      The owning entity which manages this effect
  */
+import {calcEffectRemainingRounds, calcEffectRemainingSeconds} from "../combat/combat";
+import {i18n} from "../utils/utils";
+
 export function onManageActiveEffect(event, owner) {
   event.preventDefault()
   const a = event.currentTarget
@@ -66,8 +69,22 @@ export function prepareActiveEffectCategories(effects, showCreateButtons = false
     },
   }
 
-  // Iterate over active effects, classifying them into categories
+  // Iterate over active effects, classifying them into categories.
   for (let e of effects) {
+    // Also set the 'remaining time' in seconds or rounds depending on if in combat
+    if (e.isTemporary && (e.duration.seconds || e.duration.rounds)) {
+      if (game.combat) {
+        const r = calcEffectRemainingRounds(e, game.combat.round)
+        e.dlRemaining = `${r} ${Math.abs(r) > 1 ? i18n("COMBAT.Rounds") : i18n("COMBAT.Round")}`
+      } else {
+        const r = calcEffectRemainingSeconds(e, game.time.worldTime)
+        e.dlRemaining = `${r} ${i18n("TIME.Seconds")}`
+      }
+    } else {
+      e.dlRemaining = e.duration.label
+    }
+
+
     if (e.disabled) categories.inactive.effects.push(e)
     else if (e.isTemporary) categories.temporary.effects.push(e)
     else categories.passive.effects.push(e)